--HG--
branch : spec_SFEperl-class-accessor-lite

+++ specs/SFEperl-net-server.spec
@@ -1,19 +1,19 @@
 %include Solaris.inc
 
 %define build584 0
-# build510 is for OpenIndiana,
 %define build510 %( if [ -x /usr/perl5/5.10/bin/perl ]; then echo '1'; else echo '0'; fi)
 %define build512 %( if [ -x /usr/perl5/5.12/bin/perl ]; then echo '1'; else echo '0'; fi)
 %define build516 %( if [ -x /usr/perl5/5.16/bin/perl ]; then echo '1'; else echo '0'; fi)
-%define build520 %( if [ -x /usr/perl5/5.20/bin/perl ]; then echo '1'; else echo '0'; fi)
+%define build522 %( if [ -x /usr/perl5/5.22/bin/perl ]; then echo '1'; else echo '0'; fi)
 %define include_executable 1
 
 %define cpan_name Net-Server
 %define sfe_cpan_name net-server
+%define ips_cpan_name net-server
 
 Summary:               Extensible, general Perl server engine
 Name:                  SFEperl-%{sfe_cpan_name}
-IPS_package_name:      library/perl-5/%{sfe_cpan_name}
+IPS_package_name:      library/perl-5/%{ips_cpan_name}
 Version:               2.008
 IPS_component_version: 2.8
 License:               unknown
@@ -26,13 +26,16 @@ Extensible, general Perl server engine
 
 %if %{build584}
 %package 584
-IPS_package_name: library/perl-5/%{sfe_cpan_name}-584
+IPS_package_name: library/perl-5/%{ips_cpan_name}-584
 Summary:          Extensible, general Perl server engine
 BuildRequires:    runtime/perl-584 = *
 BuildRequires:    library/perl-5/extutils-makemaker-584
+BuildRequires:    library/perl-5/io-584
+BuildRequires:    library/perl-5/socket-584
+BuildRequires:    library/perl-5/time-hires-584
 Requires:         runtime/perl-584 = *
-Requires:         library/perl-5/io-socket-584
-Requires:         library/perl-5/posix-584
+Requires:         library/perl-5/%{ips_cpan_name}
+Requires:         library/perl-5/io-584
 Requires:         library/perl-5/socket-584
 Requires:         library/perl-5/time-hires-584
 
@@ -42,13 +45,16 @@ Extensible, general Perl server engine
 
 %if %{build510}
 %package 510
-IPS_package_name: library/perl-5/%{sfe_cpan_name}-510
+IPS_package_name: library/perl-5/%{ips_cpan_name}-510
 Summary:          Extensible, general Perl server engine
 BuildRequires:    runtime/perl-510 = *
 BuildRequires:    library/perl-5/extutils-makemaker-510
+BuildRequires:    library/perl-5/io-510
+BuildRequires:    library/perl-5/socket-510
+BuildRequires:    library/perl-5/time-hires-510
 Requires:         runtime/perl-510 = *
-Requires:         library/perl-5/io-socket-510
-Requires:         library/perl-5/posix-510
+Requires:         library/perl-5/%{ips_cpan_name}
+Requires:         library/perl-5/io-510
 Requires:         library/perl-5/socket-510
 Requires:         library/perl-5/time-hires-510
 
@@ -58,13 +64,16 @@ Extensible, general Perl server engine
 
 %if %{build512}
 %package 512
-IPS_package_name: library/perl-5/%{sfe_cpan_name}-512
+IPS_package_name: library/perl-5/%{ips_cpan_name}-512
 Summary:          Extensible, general Perl server engine
 BuildRequires:    runtime/perl-512 = *
 BuildRequires:    library/perl-5/extutils-makemaker-512
+BuildRequires:    library/perl-5/io-512
+BuildRequires:    library/perl-5/socket-512
+BuildRequires:    library/perl-5/time-hires-512
 Requires:         runtime/perl-512 = *
-Requires:         library/perl-5/io-socket-512
-Requires:         library/perl-5/posix-512
+Requires:         library/perl-5/%{ips_cpan_name}
+Requires:         library/perl-5/io-512
 Requires:         library/perl-5/socket-512
 Requires:         library/perl-5/time-hires-512
 
@@ -74,13 +83,17 @@ Extensible, general Perl server engine
 
 %if %{build516}
 %package 516
-IPS_package_name: library/perl-5/%{sfe_cpan_name}-516
+IPS_package_name: library/perl-5/%{ips_cpan_name}-516
 Summary:          Extensible, general Perl server engine
 BuildRequires:    runtime/perl-516 = *
 BuildRequires:    library/perl-5/extutils-makemaker-516
+Requires:         library/perl-5/%{ips_cpan_name}
+BuildRequires:    library/perl-5/io-516
+BuildRequires:    library/perl-5/socket-516
+BuildRequires:    library/perl-5/time-hires-516
 Requires:         runtime/perl-516 = *
-Requires:         library/perl-5/io-socket-516
-Requires:         library/perl-5/posix-516
+Requires:         library/perl-5/%{ips_cpan_name}
+Requires:         library/perl-5/io-516
 Requires:         library/perl-5/socket-516
 Requires:         library/perl-5/time-hires-516
 
@@ -88,26 +101,29 @@ Requires:         library/perl-5/time-hires-516
 Extensible, general Perl server engine
 %endif
 
-%if %{build520}
-%package 520
-IPS_package_name: library/perl-5/%{sfe_cpan_name}-520
+%if %{build522}
+%package 522
+IPS_package_name: library/perl-5/%{ips_cpan_name}-522
 Summary:          Extensible, general Perl server engine
-BuildRequires:    runtime/perl-520 = *
-BuildRequires:    library/perl-5/extutils-makemaker-520
-Requires:         runtime/perl-520 = *
-Requires:         library/perl-5/io-socket-520
-Requires:         library/perl-5/posix-520
-Requires:         library/perl-5/socket-520
-Requires:         library/perl-5/time-hires-520
+BuildRequires:    runtime/perl-522 = *
+BuildRequires:    library/perl-5/extutils-makemaker-522
+BuildRequires:    library/perl-5/io-522
+BuildRequires:    library/perl-5/socket-522
+BuildRequires:    library/perl-5/time-hires-522
+Requires:         runtime/perl-522 = *
+Requires:         library/perl-5/%{ips_cpan_name}
+Requires:         library/perl-5/io-522
+Requires:         library/perl-5/socket-522
+Requires:         library/perl-5/time-hires-522
 
-%description 520
+%description 522
 Extensible, general Perl server engine
 %endif
 
 
 %prep
 %setup -q -n %{cpan_name}-%{version}
-rm -rf %{buildroot}
+[ -d %{buildroot} ] && rm -rf %{buildroot}
 
 %build
 build_with_makefile.pl_for() {
@@ -120,8 +136,17 @@ build_with_makefile.pl_for() {
     ${bindir}/perl Makefile.PL PREFIX=%{_prefix} \
                    DESTDIR=$RPM_BUILD_ROOT \
                    LIB=${vendor_dir}
-    make
-    [ ${test} == 'without_test' ] || make test
+
+    echo ${perl_ver} | egrep '5\.(84|12)' > /dev/null
+    if [ $? -eq 0 ]
+    then
+        make CC='cc -m32' LD='cc -m32'
+        [ "x${PERL_DISABLE_TEST}" = 'xtrue' ] || [ "x${test}" = 'xwithout_test' ] || make test CC='cc -m32' LD='cc -m32'
+    else
+        make CC='cc -m64' LD='cc -m64'
+        [ "x${PERL_DISABLE_TEST}" = 'xtrue' ] || [ "x${test}" = 'xwithout_test' ] || make test CC='cc -m64' LD='cc -m64'
+    fi
+
     make pure_install
 }
 
@@ -136,29 +161,60 @@ build_with_build.pl_for() {
                    --installdirs vendor \
                    --destdir $RPM_BUILD_ROOT
     ${bindir}/perl ./Build
-    [ ${test} == 'without_test' ] || ${bindir}/perl ./Build test
+    [ "x${PERL_DISABLE_TEST}" = 'xtrue' ] || [ "x${test}" = 'xwithout_test' ] || ${bindir}/perl ./Build test
     ${bindir}/perl ./Build install --destdir $RPM_BUILD_ROOT
+    ${bindir}/perl ./Build clean
 }
 
 modify_bin_dir() {
-  perl_ver=$1
-  if [ -d $RPM_BUILD_ROOT/usr/bin ]
-  then
-    [ -d $RPM_BUILD_ROOT/usr/perl5/${perl_ver} ] || mkdir -p $RPM_BUILD_ROOT/usr/perl5/${perl_ver}
-    mv $RPM_BUILD_ROOT/usr/bin $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/bin
-  fi
+    perl_ver=$1
+    if [ -d $RPM_BUILD_ROOT/usr/bin ]
+    then
+      [ -d $RPM_BUILD_ROOT/usr/perl5/${perl_ver} ] || mkdir -p $RPM_BUILD_ROOT/usr/perl5/${perl_ver}
+      mv $RPM_BUILD_ROOT/usr/bin $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/bin
+    fi
+      
+    if [ -d $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/bin ]
+    then
+        for i in $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/bin/*
+        do
+            sed -i.bak -e "s!/usr/bin/env perl!/usr/perl5/${perl-ver}/bin/perl!" ${i}
+            [ -f ${i}.bak] || rm -f ${i}.bak
+        done
+    fi
+}
+
+modify_man_dir() {
+    perl_ver=$1
+    if [ -d $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/man ]
+    then
+        if [ -d $RPM_BUILD_ROOT%{_datadir}/man ]
+        then
+            rm -rf $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/man
+        else
+            mkdir -p $RPM_BUILD_ROOT%{_datadir}
+            mv $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/man $RPM_BUILD_ROOT%{_datadir}/
+            rm -rf $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/man
+        fi
+        if [ %{include_executable} -eq 0 ]
+        then
+            rmdir $RPM_BUILD_ROOT/usr/perl5/${perl_ver}
+        fi
+
+    fi
 }
 
 build_for() {
-  if [ -f Makefile.PL ];
-  then
-    build_with_makefile.pl_for $*
-  elif [ -f Build.PL ];
+  if [ -f Build.PL ];
   then
     build_with_build.pl_for $*
+  elif [ -f Makefile.PL ];
+  then
+    build_with_makefile.pl_for $*
   fi
 
-    modify_bin_dir $*
+  modify_bin_dir $*
+  modify_man_dir $*
 }
 
 # To build without test, pass 'without_test' to build_for commaond.
@@ -179,14 +235,14 @@ build_for 5.12
 build_for 5.16
 %endif
 
-%if %{build520}
-build_for 5.20
+%if %{build522}
+build_for 5.22
 %endif
 
 %install
-mkdir -p $RPM_BUILD_ROOT%{_datadir}
 if [ -d $RPM_BUILD_ROOT%{_prefix}/man ]
 then
+    mkdir -p $RPM_BUILD_ROOT%{_datadir}
     mv $RPM_BUILD_ROOT%{_prefix}/man $RPM_BUILD_ROOT%{_datadir}
 fi
 if [ -d $RPM_BUILD_ROOT%{_datadir}/man/man3 ]
@@ -217,7 +273,7 @@ rm -rf %{buildroot}
 %dir %attr (0755, root, sys) /usr
 /usr/perl5/vendor_perl/5.10
 %if %{include_executable}
-/usr/perl5/5.10
+/usr/perl5/5.1.0
 %endif
 %endif
 
@@ -241,18 +297,21 @@ rm -rf %{buildroot}
 %endif
 %endif
 
-%if %{build520}
-%files 520
+%if %{build522}
+%files 522
 %defattr(0755,root,bin,-)
 %dir %attr (0755, root, sys) /usr
-/usr/perl5/vendor_perl/5.20
+/usr/perl5/vendor_perl/5.22
 %if %{include_executable}
-/usr/perl5/5.20
+/usr/perl5/5.22
 %endif
 %endif
-
 
 %changelog
+* Thu Apr 06 2017 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
+- fix sed in modify_bin_dir
+* Wed Apr 05 2017 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
+- package for perl-522 is added and for perl-520 is obsolete
 * Sat Nov 07 2015 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
 - build packages for perl-516 and perl-520
 * Sun Dec 07 2014 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
