+++ specs/SFEperl-net-snmp.spec
@@ -1,17 +1,20 @@
 %include Solaris.inc
 
 %define build584 0
+%define build510 %( if [ -x /usr/perl5/5.10/bin/perl ]; then echo '1'; else echo '0'; fi)
 %define build512 %( if [ -x /usr/perl5/5.12/bin/perl ]; then echo '1'; else echo '0'; fi)
 %define build516 %( if [ -x /usr/perl5/5.16/bin/perl ]; then echo '1'; else echo '0'; fi)
-%define build520 %( if [ -x /usr/perl5/5.20/bin/perl ]; then echo '1'; else echo '0'; fi)
+%define build522 %( if [ -x /usr/perl5/5.22/bin/perl ]; then echo '1'; else echo '0'; fi)
+%define enable_test %( if [ "x${PERL_DISABLE_TEST}" = 'xtrue' ]; then echo '0'; else echo '1'; fi )
 %define include_executable 1
 
 %define cpan_name Net-SNMP
 %define sfe_cpan_name net-snmp
+%define ips_cpan_name net-snmp
 
 Summary:               Object oriented interface to SNMP
 Name:                  SFEperl-%{sfe_cpan_name}
-IPS_package_name:      library/perl-5/%{sfe_cpan_name}
+IPS_package_name:      library/perl-5/%{ips_cpan_name}
 Version:               6.0.1
 IPS_component_version: 6.0.1
 License:               perl_5
@@ -24,11 +27,24 @@ Object oriented interface to SNMP
 
 %if %{build584}
 %package 584
-IPS_package_name: library/perl-5/%{sfe_cpan_name}-584
+IPS_package_name: library/perl-5/%{ips_cpan_name}-584
 Summary:          Object oriented interface to SNMP
 BuildRequires:    runtime/perl-584 = *
 BuildRequires:    library/perl-5/module-build-584
+%if %{enable_test}
+BuildRequires:    library/perl-5/carp-584
+BuildRequires:    library/perl-5/crypt-des-584
+BuildRequires:    library/perl-5/crypt-rijndael-584
+BuildRequires:    library/perl-5/digest-hmac-584
+BuildRequires:    library/perl-5/digest-md5-584
+BuildRequires:    library/perl-5/digest-sha1-584
+BuildRequires:    library/perl-5/exporter-584
+BuildRequires:    library/perl-5/io-584
+BuildRequires:    library/perl-5/math-bigint-584
+BuildRequires:    library/perl-5/socket6-584
+%endif
 Requires:         runtime/perl-584 = *
+Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/carp-584
 Requires:         library/perl-5/crypt-des-584
 Requires:         library/perl-5/crypt-rijndael-584
@@ -36,7 +52,7 @@ Requires:         library/perl-5/digest-hmac-584
 Requires:         library/perl-5/digest-md5-584
 Requires:         library/perl-5/digest-sha1-584
 Requires:         library/perl-5/exporter-584
-Requires:         library/perl-5/io-socket-584
+Requires:         library/perl-5/io-584
 Requires:         library/perl-5/math-bigint-584
 Requires:         library/perl-5/socket6-584
 
@@ -44,13 +60,59 @@ Requires:         library/perl-5/socket6-584
 Object oriented interface to SNMP
 %endif
 
+%if %{build510}
+%package 510
+IPS_package_name: library/perl-5/%{ips_cpan_name}-510
+Summary:          Object oriented interface to SNMP
+BuildRequires:    runtime/perl-510 = *
+BuildRequires:    library/perl-5/module-build-510
+BuildRequires:    library/perl-5/carp-510
+BuildRequires:    library/perl-5/crypt-des-510
+BuildRequires:    library/perl-5/crypt-rijndael-510
+BuildRequires:    library/perl-5/digest-hmac-510
+BuildRequires:    library/perl-5/digest-md5-510
+BuildRequires:    library/perl-5/digest-sha1-510
+BuildRequires:    library/perl-5/exporter-510
+BuildRequires:    library/perl-5/io-510
+BuildRequires:    library/perl-5/math-bigint-510
+BuildRequires:    library/perl-5/socket6-510
+Requires:         runtime/perl-510 = *
+Requires:         library/perl-5/%{ips_cpan_name}
+Requires:         library/perl-5/carp-510
+Requires:         library/perl-5/crypt-des-510
+Requires:         library/perl-5/crypt-rijndael-510
+Requires:         library/perl-5/digest-hmac-510
+Requires:         library/perl-5/digest-md5-510
+Requires:         library/perl-5/digest-sha1-510
+Requires:         library/perl-5/exporter-510
+Requires:         library/perl-5/io-510
+Requires:         library/perl-5/math-bigint-510
+Requires:         library/perl-5/socket6-510
+
+%description 510
+Object oriented interface to SNMP
+%endif
+
 %if %{build512}
 %package 512
-IPS_package_name: library/perl-5/%{sfe_cpan_name}-512
+IPS_package_name: library/perl-5/%{ips_cpan_name}-512
 Summary:          Object oriented interface to SNMP
 BuildRequires:    runtime/perl-512 = *
 BuildRequires:    library/perl-5/module-build-512
+%if %{enable_test}
+BuildRequires:    library/perl-5/carp-512
+BuildRequires:    library/perl-5/crypt-des-512
+BuildRequires:    library/perl-5/crypt-rijndael-512
+BuildRequires:    library/perl-5/digest-hmac-512
+BuildRequires:    library/perl-5/digest-md5-512
+BuildRequires:    library/perl-5/digest-sha1-512
+BuildRequires:    library/perl-5/exporter-512
+BuildRequires:    library/perl-5/io-512
+BuildRequires:    library/perl-5/math-bigint-512
+BuildRequires:    library/perl-5/socket6-512
+%endif
 Requires:         runtime/perl-512 = *
+Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/carp-512
 Requires:         library/perl-5/crypt-des-512
 Requires:         library/perl-5/crypt-rijndael-512
@@ -58,7 +120,7 @@ Requires:         library/perl-5/digest-hmac-512
 Requires:         library/perl-5/digest-md5-512
 Requires:         library/perl-5/digest-sha1-512
 Requires:         library/perl-5/exporter-512
-Requires:         library/perl-5/io-socket-512
+Requires:         library/perl-5/io-512
 Requires:         library/perl-5/math-bigint-512
 Requires:         library/perl-5/socket6-512
 
@@ -68,11 +130,25 @@ Object oriented interface to SNMP
 
 %if %{build516}
 %package 516
-IPS_package_name: library/perl-5/%{sfe_cpan_name}-516
+IPS_package_name: library/perl-5/%{ips_cpan_name}-516
 Summary:          Object oriented interface to SNMP
 BuildRequires:    runtime/perl-516 = *
 BuildRequires:    library/perl-5/module-build-516
+Requires:         library/perl-5/%{ips_cpan_name}
+%if %{enable_test}
+BuildRequires:    library/perl-5/carp-516
+BuildRequires:    library/perl-5/crypt-des-516
+BuildRequires:    library/perl-5/crypt-rijndael-516
+BuildRequires:    library/perl-5/digest-hmac-516
+BuildRequires:    library/perl-5/digest-md5-516
+BuildRequires:    library/perl-5/digest-sha1-516
+BuildRequires:    library/perl-5/exporter-516
+BuildRequires:    library/perl-5/io-516
+BuildRequires:    library/perl-5/math-bigint-516
+BuildRequires:    library/perl-5/socket6-516
+%endif
 Requires:         runtime/perl-516 = *
+Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/carp-516
 Requires:         library/perl-5/crypt-des-516
 Requires:         library/perl-5/crypt-rijndael-516
@@ -80,7 +156,7 @@ Requires:         library/perl-5/digest-hmac-516
 Requires:         library/perl-5/digest-md5-516
 Requires:         library/perl-5/digest-sha1-516
 Requires:         library/perl-5/exporter-516
-Requires:         library/perl-5/io-socket-516
+Requires:         library/perl-5/io-516
 Requires:         library/perl-5/math-bigint-516
 Requires:         library/perl-5/socket6-516
 
@@ -88,32 +164,45 @@ Requires:         library/perl-5/socket6-516
 Object oriented interface to SNMP
 %endif
 
-%if %{build520}
-%package 520
-IPS_package_name: library/perl-5/%{sfe_cpan_name}-520
+%if %{build522}
+%package 522
+IPS_package_name: library/perl-5/%{ips_cpan_name}-522
 Summary:          Object oriented interface to SNMP
-BuildRequires:    runtime/perl-520 = *
-BuildRequires:    library/perl-5/module-build-520
-Requires:         runtime/perl-520 = *
-Requires:         library/perl-5/carp-520
-Requires:         library/perl-5/crypt-des-520
-Requires:         library/perl-5/crypt-rijndael-520
-Requires:         library/perl-5/digest-hmac-520
-Requires:         library/perl-5/digest-md5-520
-Requires:         library/perl-5/digest-sha1-520
-Requires:         library/perl-5/exporter-520
-Requires:         library/perl-5/io-socket-520
-Requires:         library/perl-5/math-bigint-520
-Requires:         library/perl-5/socket6-520
+BuildRequires:    runtime/perl-522 = *
+BuildRequires:    library/perl-5/module-build-522
+%if %{enable_test}
+BuildRequires:    library/perl-5/carp-522
+BuildRequires:    library/perl-5/crypt-des-522
+BuildRequires:    library/perl-5/crypt-rijndael-522
+BuildRequires:    library/perl-5/digest-hmac-522
+BuildRequires:    library/perl-5/digest-md5-522
+BuildRequires:    library/perl-5/digest-sha1-522
+BuildRequires:    library/perl-5/exporter-522
+BuildRequires:    library/perl-5/io-522
+BuildRequires:    library/perl-5/math-bigint-522
+BuildRequires:    library/perl-5/socket6-522
+%endif
+Requires:         runtime/perl-522 = *
+Requires:         library/perl-5/%{ips_cpan_name}
+Requires:         library/perl-5/carp-522
+Requires:         library/perl-5/crypt-des-522
+Requires:         library/perl-5/crypt-rijndael-522
+Requires:         library/perl-5/digest-hmac-522
+Requires:         library/perl-5/digest-md5-522
+Requires:         library/perl-5/digest-sha1-522
+Requires:         library/perl-5/exporter-522
+Requires:         library/perl-5/io-522
+Requires:         library/perl-5/math-bigint-522
+Requires:         library/perl-5/socket6-522
 
-%description 520
+%description 522
 Object oriented interface to SNMP
 %endif
 
 
 %prep
 %setup -q -n %{cpan_name}-v%{version}
-rm -rf %{buildroot}
+[ -d %{buildroot} ] && rm -rf %{buildroot}
 
 %build
 build_with_makefile.pl_for() {
@@ -126,8 +215,12 @@ build_with_makefile.pl_for() {
     ${bindir}/perl Makefile.PL PREFIX=%{_prefix} \
                    DESTDIR=$RPM_BUILD_ROOT \
                    LIB=${vendor_dir}
-    make
-    [ ${test} = 'without_test' ] || make test
+
+    export CC='cc -m32'
+    export LD='cc -m32'
+    echo ${perl_ver} | egrep '5\.(84|12)' > /dev/null || (export CC='cc -m64'; export LD='cc -m64')
+    make CC="${CC}" LD="${LD}"
+    [ "x${PERL_DISABLE_TEST}" = 'xtrue' ] || [ "x${test}" = 'xwithout_test' ] || make test CC="${CC}" "LD=${LD}"
     make pure_install
 }
 
@@ -142,32 +235,47 @@ build_with_build.pl_for() {
                    --installdirs vendor \
                    --destdir $RPM_BUILD_ROOT
     ${bindir}/perl ./Build
-    [ ${test} = 'without_test' ] || ${bindir}/perl ./Build test
+    [ "x${PERL_DISABLE_TEST}" = 'xtrue' ] || [ "x${test}" = 'xwithout_test' ] || ${bindir}/perl ./Build test
     ${bindir}/perl ./Build install --destdir $RPM_BUILD_ROOT
+    ${bindir}/perl ./Build clean
 }
 
 modify_bin_dir() {
-  perl_ver=$1
-  if [ -d $RPM_BUILD_ROOT/usr/bin ]
-  then
-    [ -d $RPM_BUILD_ROOT/usr/perl5/${perl_ver} ] || mkdir -p $RPM_BUILD_ROOT/usr/perl5/${perl_ver}
-    mv $RPM_BUILD_ROOT/usr/bin $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/bin
-  fi
+    perl_ver=$1
+    if [ -d $RPM_BUILD_ROOT/usr/bin ]
+    then
+      [ -d $RPM_BUILD_ROOT/usr/perl5/${perl_ver} ] || mkdir -p $RPM_BUILD_ROOT/usr/perl5/${perl_ver}
+      mv $RPM_BUILD_ROOT/usr/bin $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/bin
+    fi
+      
+    if [ -d $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/bin ]
+    then
+        for i in $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/bin/*
+        do
+            sed -i.bak -e "s!/usr/bin/env perl!/usr/perl5/${perl-ver}/bin/perl!" ${i}
+            [ -f ${i}.bak] || rm -f ${i}.bak
+        done
+    fi
 }
 
 modify_man_dir() {
-  perl_ver=$1
-  if [ -d $RPM_BUILD_ROOT/usr/perl/${perl_ver}/man ]
-  then
-      if [ -d $RPM_BUILD_ROOT%{_datadir}/man ]
-      then
-	  mkdir -p $RPM_BUILD_ROOT%{_datadir}/man
-	  mv $RPM_BUILD_ROOT/usr/perl/${perl_ver}/man/* $RPM_BUILD_ROOT%{_datadir}/man
-      else
-	  rm -rf $RPM_BUILD_ROOT%{_datadir}/man
-      fi
-  fi
+    perl_ver=$1
+    if [ -d $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/man ]
+    then
+        if [ -d $RPM_BUILD_ROOT%{_datadir}/man ]
+        then
+            rm -rf $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/man
+        else
+            mkdir -p $RPM_BUILD_ROOT%{_datadir}
+            mv $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/man $RPM_BUILD_ROOT%{_datadir}/
+            rm -rf $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/man
+        fi
+        if [ %{include_executable} -eq 0 ]
+        then
+            rmdir $RPM_BUILD_ROOT/usr/perl5/${perl_ver}
+        fi
 
+    fi
 }
 
 build_for() {
@@ -189,6 +297,10 @@ build_for() {
 build_for 5.8.4
 %endif
 
+%if %{build510}
+build_for 5.10
+%endif
+
 %if %{build512}
 build_for 5.12
 %endif
@@ -197,26 +309,27 @@ build_for 5.12
 build_for 5.16
 %endif
 
-%if %{build520}
-build_for 5.20
+%if %{build522}
+build_for 5.22
 %endif
 
 %install
-# mkdir -p $RPM_BUILD_ROOT%{_datadir}
-# if [ -d $RPM_BUILD_ROOT%{_prefix}/man ]
-# then
-#     mv $RPM_BUILD_ROOT%{_prefix}/man $RPM_BUILD_ROOT%{_datadir}
-# fi
-# if [ -d $RPM_BUILD_ROOT%{_datadir}/man/man3 ]
-# then
-#     mv $RPM_BUILD_ROOT%{_datadir}/man/man3 $RPM_BUILD_ROOT%{_datadir}/man/man3perl
-# fi
+if [ -d $RPM_BUILD_ROOT%{_prefix}/man ]
+then
+    mkdir -p $RPM_BUILD_ROOT%{_datadir}
+    mv $RPM_BUILD_ROOT%{_prefix}/man $RPM_BUILD_ROOT%{_datadir}
+fi
+if [ -d $RPM_BUILD_ROOT%{_datadir}/man/man3 ]
+then
+    mv $RPM_BUILD_ROOT%{_datadir}/man/man3 $RPM_BUILD_ROOT%{_datadir}/man/man3perl
+fi
 
 %clean
 rm -rf %{buildroot}
 
 %files
 %defattr(0755,root,bin,-)
+%{_datadir}/man
 
 %if %{build584}
 %files 584
@@ -225,6 +338,16 @@ rm -rf %{buildroot}
 /usr/perl5/vendor_perl/5.8.4
 %if %{include_executable}
 /usr/perl5/5.8.4
+%endif
+%endif
+
+%if %{build510}
+%files 510
+%defattr(0755,root,bin,-)
+%dir %attr (0755, root, sys) /usr
+/usr/perl5/vendor_perl/5.10
+%if %{include_executable}
+/usr/perl5/5.1.0
 %endif
 %endif
 
@@ -248,18 +371,23 @@ rm -rf %{buildroot}
 %endif
 %endif
 
-%if %{build520}
-%files 520
+%if %{build522}
+%files 522
 %defattr(0755,root,bin,-)
 %dir %attr (0755, root, sys) /usr
-/usr/perl5/vendor_perl/5.20
+/usr/perl5/vendor_perl/5.22
 %if %{include_executable}
-/usr/perl5/5.20
+/usr/perl5/5.22
 %endif
 %endif
-
 
 %changelog
+* Thu Apr 27 2017 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
+- fix build
+* Tue Apr 04 2017 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
+- fix sed in modify_bin_dir
+* Wed Apr 05 2017 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
+- package for perl-522 is added and for perl-520 is obsolete
 * Wed Nov 11 2015 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
 - build packages for perl-516 and perl-520
 * Mon Feb 11 2013 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
